pyproject.toml
[project]
name = "customer-support-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
